refactor: extract pattern list
