+++ src/components/index.js
@@ -8,3 +8,4 @@ export { default as Logo } from './logo';
 export { default as ContentLoader } from './content-loader';
 export { default as ScrollToTop } from './scroll-to-top';
 export { default as CategoryCard } from './category-card';
+export { default as PatternList } from './pattern-list';
